Feature issue#3 Integration test worker
Added an integration test worker that subscribes to the deploy queue, and publishes
to the after_deploy, regardless of the outcome.

Currently, no database writing is going on, assuming there will be a worker doing
this task.

Tests have been included for functional tests - although they are not THAT useful,
as the core of the test is mocked. More educational than anything.

Unit tests should cover most of the scenarios expected from raising, to failing,
and cleanup should be carried out.

TODO: adsrex configuration on runtime