Issue #6: Database worker
Preliminary idea of what we would store for each transaction that occurs for a worker.
The logic can exist within the single database worker, or it could exist inside that of
each worker in the queue.

This does require that each worker pass on the relevant information to the next worker.
For example, there should be values in the payload like 'deploy' from start to finish,
so that the relevant value should be updated if necessary.