Feature Issue#17: queue forwarding
Replaced the inefficient call of .forward() with .publish(), and
used a configuration parameter to define the status queue, much
like the publish_to_error_queue works. Although it is a tiny bit
more boiler plate code, given (nearly) all workers publish to the
status queue, it is neater to see it in the config.

Updated the tests for changes made.

Added functional tests for the Restart worker, that are passing.

Fixed the broken tests.

Currently, version == commit. This should be changed in the future.